+++ src/components/Catalogs.vue
@@ -7,25 +7,32 @@
       <SortButtons v-if="!hideControls && isComplete && catalogs.length > 1" v-model="sort.direction" />
     </header>
     <section v-if="!hideControls && ((isComplete && catalogs.length > 1) || canSearchFreeText)" class="catalog-filter mb-2">
-      <multiselect
-        v-if="canSearchFreeText" multiple taggable @tag="addSearchTerm"
-        id="catalogFreeText" v-model="selectedSearchTerms" :options="selectedSearchTerms"
-        :placeholder="$t('search.enterSearchTerms')" :tag-placeholder="$t('search.addSearchTerm')" :no-options="$t('search.addSearchTerm')"
-      >
-        <template #noOptions>{{ $t('search.noOptions') }}</template>
-      </multiselect>
-      <SearchBox v-else v-model="searchTerm" :placeholder="filterPlaceholder" />
-      <multiselect
-        v-if="isComplete && allKeywords.length > 0"
-        v-model="selectedKeywords"
-        :options="allKeywords"
-        multiple
-        :placeholder="$t('multiselect.keywordsPlaceholder')"
-        :select-label="$t('multiselect.selectLabel')"
-        :selected-label="$t('multiselect.selectedLabel')"
-        :deselect-label="$t('multiselect.deselectLabel')"
-        :limit-text="limitText"
-      />
+      <template v-if="canSearchFreeText">
+        <multiselect
+          multiple taggable @tag="addSearchTerm"
+          id="catalogFreeText" v-model="selectedSearchTerms" :options="selectedSearchTerms"
+          :placeholder="$t('search.enterSearchTerms')" :tag-placeholder="$t('search.addSearchTerm')" :no-options="$t('search.addSearchTerm')"
+        >
+          <template #noOptions>{{ $t('search.noOptions') }}</template>
+        </multiselect>
+        <b-button v-if="advancedSearchLink" variant="primary" :to="advancedSearchLink" class="additional-filter-link">
+          {{ $t('search.additionalFilters') }}
+        </b-button>
+      </template>
+      <template v-else>
+        <SearchBox v-model="searchTerm" :placeholder="filterPlaceholder" />
+        <multiselect
+          v-if="isComplete && allKeywords.length > 0"
+          v-model="selectedKeywords"
+          :options="allKeywords"
+          multiple
+          :placeholder="$t('multiselect.keywordsPlaceholder')"
+          :select-label="$t('multiselect.selectLabel')"
+          :selected-label="$t('multiselect.selectedLabel')"
+          :deselect-label="$t('multiselect.deselectLabel')"
+          :limit-text="limitText"
+        />
+      </template>
     </section>
     <Pagination v-if="showPagination" ref="topPagination" class="mb-3" :pagination="pagination" placement="top" @paginate="paginate" />
     <b-alert v-if="hasSearchCritera && catalogView.length === 0" variant="warning" class="mt-2" show>{{ $t('catalogs.noMatches') }}</b-alert>
@@ -58,7 +65,7 @@ import ViewButtons from './ViewButtons.vue';
 import Utils from '../utils';
 import { sortStac } from '../models/stac';
 import { TYPES } from './ApiCapabilitiesMixin.js';
-import { hasText } from 'stac-js/src/utils.js';
+import { hasText, URI } from 'stac-js/src/utils.js';
 
 export default defineComponent({
   name: "Catalogs",
@@ -128,8 +135,18 @@ export default defineComponent({
   },
   computed: {
     ...mapState(['defaultCollectionSort', 'uiLanguage']),
-    ...mapGetters(['supportsConformance']),
+    ...mapGetters(['searchBrowserLink', 'supportsConformance']),
     ...mapGetters(['getStac']),
+    advancedSearchLink() {
+      if (!this.canSearchFreeText || !this.searchBrowserLink) {
+        return null;
+      }
+      const query = Utils.stateQueryParametersToObject({
+        'searchtype': 'collections',
+        'q': this.selectedSearchTerms,
+      });
+      return URI(this.searchBrowserLink).query(query).toString();
+    },
     canSearchFreeText() {
       return this.apiSearch && this.supportsConformance(TYPES.Collections.FreeText);
     },
@@ -291,6 +308,12 @@ export default defineComponent({
     flex-grow: 1;
     flex-basis: 300px;
     min-width: 300px;
+  }
+
+  > .additional-filter-link {
+    flex-grow: 0;
+    align-self: center;
+    white-space: nowrap;
   }
 }
 </style>